# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APS - Automotive Platform System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partner/jobs:
    post:
      operationId: createPartnerJob
      tags: [jobs]
      summary: Post a job from a Fleet or Dealership partner
      description: |
        Fleet & Commercial Partner job posting. Only owners of shops with
        `partnerKind ∈ {dealership, fleet, gsa}` may post here — independent
        shops rent bays via `/bays` instead. The server stamps a flat
        commission override (default 15%, or 10% for GSA accounts when
        `shop.commissionOverridePct` is configured) and a tier-priority
        window: `urgent`=visible to all qualified mechanics immediately,
        `high`=junior tiers wait 15m, `normal`=1h, `low`=4h. Senior+
        mechanics always see partner-posted jobs immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                type: object
                required: [id, jobId, urgency, juniorVisibleAt, commissionPctOverride]
                properties:
                  id: { type: integer }
                  jobId: { type: integer }
                  urgency: { type: string, enum: [low, normal, high, urgent] }
                  juniorVisibleAt: { type: string, format: date-time }
                  commissionPctOverride: { type: integer }
        "403":
          description: Not a fleet/dealership partner, or not the shop owner
...
  /shops:
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a new shop (shop_owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

  # ────────────── Partner organizations ──────────────
...
  /shops/{shopId}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a shop with its bays
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopWithBays"
    patch:
      operationId: updateShop
      tags: [shops]
      summary: Update a shop (owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatePartnerJobBody:
      type: object
      required: [shopId, vehicleId, description]
      properties:
        shopId:
          type: integer
          description: Partner shop posting the job. Must have partnerKind=dealership or fleet.
        vehicleId:
          type: integer
          description: Vehicle from the partner's fleet (must be pre-registered via POST /vehicles).
        serviceSlug:
          type: string
          description: Preferred. Slug from /tier-catalog. Server derives jobType + requiredTier.
        jobType:
          type: string
          enum: [repair, diagnostic, maintenance, detailing]
          description: Optional fallback when serviceSlug is not supplied.
        description:
          type: string
        locationLat: { type: number }
        locationLng: { type: number }
        locationAddress: { type: string }
        urgency:
          type: string
          enum: [low, normal, high, urgent]
          description: Defaults to `normal`. Drives the tier-priority window for who sees the job when.
        recurringGroupId:
          type: string
          description: Optional. Stamps a group id for bulk/recurring fleet ops (e.g. monthly oil changes across 12 vehicles).
...
    Shop:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        organizationId:
          type: ["integer", "null"]
        partnerKind:
          type: string
          enum: [independent_shop, dealership, fleet, gsa]
          description: |
            Discriminates the partner type. `independent_shop` uses the classic
            bay-rental marketplace. `dealership`, `fleet`, and `gsa` (Government /
            GSA accounts) can additionally post overflow jobs at a reduced
            commission.
        name:
          type: string
        address:
          type: string
        city:
          type: string
        region:
          type: string
        zipCode:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        phone:
          type: string
          nullable: true
        federalEin:
          type: string
          nullable: true
          description: Federal EIN / Tax ID — used to uniquely identify the partner business with the IRS.
        businessLicense:
          type: string
          nullable: true
          description: State or local business license number.
        insuranceCarrier:
          type: string
          nullable: true
          deprecated: true
          description: DEPRECATED. Per-vehicle insurance now lives on `Vehicle.insuranceCarrier`.
        insurancePolicyNumber:
          type: string
          nullable: true
          deprecated: true
          description: DEPRECATED. Per-vehicle insurance now lives on `Vehicle.insurancePolicyNumber`.
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, organizationId, partnerKind, name, address, city, region, zipCode, status, createdAt]
...
    CreateShopBody:
      type: object
      properties:
        partnerKind:
          type: string
          enum: [independent_shop, dealership, fleet, gsa]
          default: independent_shop
        name:
          type: string
        address:
          type: string
        city:
          type: string
        region:
          type: string
        zipCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        phone:
          type: string
        federalEin:
          type: string
          description: Federal EIN / Tax ID.
        businessLicense:
          type: string
          description: State or local business license number.
        insuranceCarrier:
          type: string
          deprecated: true
        insurancePolicyNumber:
          type: string
          deprecated: true
      required: [name, address, city, region, zipCode]
    UpdateShopBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        region:
          type: string
        zipCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        phone:
          type: string
        federalEin:
          type: string
        businessLicense:
          type: string
        insuranceCarrier:
          type: string
          deprecated: true
        insurancePolicyNumber:
          type: string
          deprecated: true
        status:
          type: string
          enum: [active, inactive]
    Bay:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        hourlyRate:
          type: number
        equipment:
          type: array
          items:
            type: string
        allowedJobCategories:
          type: array
          items:
            type: string
            enum: [repair, diagnostic, maintenance, detailing]
        minMechanicTier:
          type: string
          enum: [detailer, technician, senior, advanced, master]
        autoApprove:
          type: boolean
        availabilityConfig:
          $ref: "#/components/schemas/BayAvailabilityConfig"
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, shopId, name, hourlyRate, equipment, allowedJobCategories, minMechanicTier, autoApprove, availabilityConfig, status, createdAt]
    ShopWithBays:
      allOf:
        - $ref: "#/components/schemas/Shop"
        - type: object
          properties:
            bays:
              type: array
              items:
                $ref: "#/components/schemas/Bay"
          required: [bays]
...
    BayAvailabilityConfig:
      type: object
      additionalProperties: false
      properties:
        timezone:
          type: string
          enum: [UTC]
        weekly:
          type: array
          items:
            $ref: "#/components/schemas/BayAvailabilityWindow"
      required: [timezone, weekly]
    BayAvailabilityWindow:
      type: object
      additionalProperties: false
      properties:
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        open:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        close:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
      required: [dayOfWeek, open, close]
...